backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["Eric Alfred <alfrederic371@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^5.2.1"
django-cors-headers = "^4.7.0"
psutil = "^5.9.8"
netifaces = "^0.11.0"
dbus-python = "^1.4.0"
psutils = "^3.3.10"
pillow = "^11.2.1"
python-xlib = "^0.33"
zeroconf = "^0.131.0"
djangorestframework = "^3.14.0"
djangorestframework-simplejwt = "^5.3.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
